Move AddTemplate to its own module
